Adapt to moving of definitions to ghc-internal
