--- src/Data/Int_LHAssumptions.hs
@@ -1,10 +1,0 @@
-{-# OPTIONS_GHC -fplugin=LiquidHaskellBoot #-}
-module Data.Int_LHAssumptions where
-
-{-@
-embed Data.Int.Int8  as int
-embed Data.Int.Int16 as int
-embed Data.Int.Int32 as int
-embed Data.Int.Int64 as int
-
-@-}
