+++ src/Data/Word_LHAssumptions.hs
@@ -1,13 +1,4 @@
 {-# OPTIONS_GHC -fplugin=LiquidHaskellBoot #-}
 module Data.Word_LHAssumptions where
 
-{-@
-embed GHC.Word.Word   as int
-embed GHC.Word.Word8  as int
-embed GHC.Word.Word16 as int
-embed GHC.Word.Word32 as int
-embed GHC.Word.Word64 as int
-
-invariant {v : GHC.Word.Word32 | 0 <= v }
-invariant {v : GHC.Word.Word16 | 0 <= v }
-@-}
+import GHC.Internal.Word_LHAssumptions()
