+++ src/GHC/Num/Integer_LHAssumptions.hs
@@ -5,9 +5,11 @@ module GHC.Num.Integer_LHAssumptions() where
 
 import GHC.Types
 import GHC.Num.Integer
+import GHC.Prim
 import GHC.Types_LHAssumptions()
-
 
 {-@
 assume GHC.Num.Integer.IS :: x:GHC.Prim.Int# -> {v: GHC.Num.Integer.Integer | v = (x :: int) }
+
+embed GHC.Num.Integer.Integer as int
 @-}
