+++ src/GHC/Word_LHAssumptions.hs
@@ -1,10 +1,4 @@
 {-# OPTIONS_GHC -fplugin=LiquidHaskellBoot #-}
 module GHC.Word_LHAssumptions where
 
-{-@
-embed GHC.Word.Word   as int
-embed GHC.Word.Word8  as int
-embed GHC.Word.Word16 as int
-embed GHC.Word.Word32 as int
-embed GHC.Word.Word64 as int
-@-}
+import GHC.Internal.Word_LHAssumptions()
