+++ src/Prelude_LHAssumptions.hs
@@ -6,6 +6,7 @@ import GHC.Float_LHAssumptions()
 import GHC.Maybe_LHAssumptions()
 import GHC.Num_LHAssumptions()
 import GHC.Num.Integer_LHAssumptions()
+import GHC.Num.Integer()
 import GHC.Real_LHAssumptions()
 import Liquid.Prelude.Real_LHAssumptions()
 import Liquid.Prelude.Totality_LHAssumptions()
