+++ tests/pos/Comprehension.hs
@@ -4,4 +4,4 @@ module Comprehension where
 foo :: Int -> [Int]
 foo n = [0 .. n]
 
-{-@ assume GHC.Enum.enumFromTo :: (Enum a) => lo:a -> hi:a -> [{v:a | lo <= v && v <= hi}] @-}
+{-@ assume GHC.Internal.Enum.enumFromTo :: (Enum a) => lo:a -> hi:a -> [{v:a | lo <= v && v <= hi}] @-}
